refactor: move SCC and cycles find algorithms to separate classes
